+++ src/main/java/com/bamdoliro/gati/domain/user/service/AuthService.java
@@ -1,0 +1,48 @@
+package com.bamdoliro.gati.domain.user.service;
+
+import com.bamdoliro.gati.domain.user.domain.User;
+import com.bamdoliro.gati.domain.user.facade.UserFacade;
+import com.bamdoliro.gati.domain.user.presentation.dto.request.LoginUserRequestDto;
+import com.bamdoliro.gati.domain.user.presentation.dto.response.TokenResponseDto;
+import com.bamdoliro.gati.global.redis.RedisService;
+import com.bamdoliro.gati.global.security.jwt.JwtTokenProvider;
+import com.bamdoliro.gati.global.security.jwt.JwtValidateService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import static com.bamdoliro.gati.global.security.jwt.JwtProperties.REFRESH_TOKEN_VALID_TIME;
+
+@Service
+@RequiredArgsConstructor
+public class AuthService {
+
+    private final JwtTokenProvider jwtTokenProvider;
+    private final JwtValidateService jwtValidateService;
+    private final UserFacade userFacade;
+    private final RedisService redisService;
+
+    public TokenResponseDto loginUser(LoginUserRequestDto dto) {
+        User user = userFacade.findUserByEmail(dto.getEmail());
+        userFacade.checkUserPassword(user, dto.getPassword());
+
+        final String accessToken = jwtTokenProvider.createAccessToken(dto.getEmail());
+        final String refreshToken = jwtTokenProvider.createRefreshToken(dto.getEmail());
+        redisService.setDataExpire(dto.getEmail(), refreshToken, REFRESH_TOKEN_VALID_TIME);
+
+        return TokenResponseDto.builder()
+                .accessToken(accessToken)
+                .refreshToken(refreshToken)
+                .build();
+    }
+
+    public void logoutUser() {
+        User user = userFacade.getCurrentUser();
+        redisService.deleteData(user.getEmail());
+    }
+
+    // TODO :: new access token with refresh token
+    private void setNewAccessToken(String refreshToken) {
+        String newAccessToken = jwtTokenProvider.createAccessToken(
+                jwtValidateService.getEmail(refreshToken));
+    }
+}
