+++ src/main/java/com/bamdoliro/gati/domain/user/service/UserService.java
@@ -3,17 +3,15 @@ package com.bamdoliro.gati.domain.user.service;
 import com.bamdoliro.gati.domain.user.domain.User;
 import com.bamdoliro.gati.domain.user.domain.repository.UserRepository;
 import com.bamdoliro.gati.domain.user.facade.UserFacade;
-import com.bamdoliro.gati.domain.user.presentation.dto.request.*;
+import com.bamdoliro.gati.domain.user.presentation.dto.request.CreateUserRequestDto;
+import com.bamdoliro.gati.domain.user.presentation.dto.request.DeleteUserRequestDto;
+import com.bamdoliro.gati.domain.user.presentation.dto.request.UpdateUserNameRequestDto;
+import com.bamdoliro.gati.domain.user.presentation.dto.request.UpdateUserPasswordRequestDto;
 import com.bamdoliro.gati.domain.user.presentation.dto.response.GetUserResponseDto;
-import com.bamdoliro.gati.domain.user.presentation.dto.response.TokenResponseDto;
-import com.bamdoliro.gati.global.redis.RedisService;
-import com.bamdoliro.gati.global.security.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import static com.bamdoliro.gati.global.security.jwt.JwtProperties.REFRESH_TOKEN_VALID_TIME;
 
 @Service
 @RequiredArgsConstructor
@@ -21,8 +19,6 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
-    private final RedisService redisService;
-    private final JwtTokenProvider jwtTokenProvider;
     private final UserFacade userFacade;
 
     @Transactional
@@ -30,29 +26,6 @@ public class UserService {
         userFacade.checkUserByEmail(dto.getEmail());
 
         userRepository.save(dto.toEntity(passwordEncoder.encode(dto.getPassword())));
-    }
-
-    public TokenResponseDto loginUser(LoginUserRequestDto dto) {
-        User user = userFacade.findUserByEmail(dto.getEmail());
-        userFacade.checkUserPassword(user, dto.getPassword());
-
-        final String accessToken = jwtTokenProvider.createAccessToken(dto.getEmail());
-        final String refreshToken = jwtTokenProvider.createRefreshToken(dto.getEmail());
-        redisService.setDataExpire(dto.getEmail(), refreshToken, REFRESH_TOKEN_VALID_TIME);
-
-        return TokenResponseDto.builder()
-                .accessToken(accessToken)
-                .refreshToken(refreshToken)
-                .build();
-    }
-
-    public void logoutUser() {
-        User user = userFacade.getCurrentUser();
-
-//        Cookie tempCookie1 = cookieUtil.deleteCookie(ACCESS_TOKEN_NAME);
-//        Cookie tempCookie2 = cookieUtil.deleteCookie(REFRESH_TOKEN_NAME);
-
-        redisService.deleteData(user.getEmail());
     }
 
     @Transactional(readOnly = true)
